requirements.txt
python-dotenv==1.2.1
python-telegram-bot==13.15
redis==7.0.1
vk_api==11.10.0
